// fct-frontend/src/api/models/error-message.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ErrorMessage
 */
export interface ErrorMessage {
    /**
     * 
     * @type {string}
     * @memberof ErrorMessage
     */
    'timestamp': string;
    /**
     * 
     * @type {number}
     * @memberof ErrorMessage
     */
    'status': number;
    /**
     * 
     * @type {string}
     * @memberof ErrorMessage
     */
    'error': string;
    /**
     * 
     * @type {string}
     * @memberof ErrorMessage
     */
    'message': string;
    /**
     * 
     * @type {string}
     * @memberof ErrorMessage
     */
    'path': string;
}