// fct-frontend/src/api/models/role-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API for the website of \"Academia Fabiana Perez\"
 * REST API for the website of \"Academia Fabiana Perez\"
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const RoleEnum = {
    ADMIN: 'ADMIN',
    USER: 'USER'
} as const;

export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];